chore: move TipTapEditor into its own folder
It's not really just used for promptResponse anymore.
