+++ packages/client/components/ReflectionCard/ReflectionCard.tsx
@@ -24,8 +24,8 @@ import {cn} from '../../ui/cn'
 import isPhaseComplete from '../../utils/meetings/isPhaseComplete'
 import isTempId from '../../utils/relay/isTempId'
 import type {OpenSpotlight} from '../GroupingKanbanColumn'
-import {TipTapEditor} from '../promptResponse/TipTapEditor'
 import StyledError from '../StyledError'
+import {TipTapEditor} from '../TipTapEditor/TipTapEditor'
 import ColorBadge from './ColorBadge'
 import DeleteReflectionButton from './DeleteReflectionButton'
 import ReactjiSection from './ReactjiSection'
