+++ packages/client/components/ResponseReplied.tsx
@@ -6,7 +6,7 @@ import useRouter from '../hooks/useRouter'
 import {useTipTapCommentEditor} from '../hooks/useTipTapCommentEditor'
 import anonymousAvatar from '../styles/theme/images/anonymous-avatar.svg'
 import NotificationTemplate from './NotificationTemplate'
-import {TipTapEditor} from './promptResponse/TipTapEditor'
+import {TipTapEditor} from './TipTapEditor/TipTapEditor'
 
 interface Props {
   notification: ResponseReplied_notification$key
