+++ packages/client/components/RetroReflectPhase/PhaseItemEditor.tsx
@@ -16,10 +16,10 @@ import {Elevation} from '../../styles/elevation'
 import {BezierCurve, ZIndex} from '../../types/constEnums'
 import {cn} from '../../ui/cn'
 import {modEnter} from '../../utils/platform'
-import {TipTapEditor} from '../promptResponse/TipTapEditor'
 import ReflectionCardAuthor from '../ReflectionCard/ReflectionCardAuthor'
 import ReflectionCardRoot from '../ReflectionCard/ReflectionCardRoot'
 import SubmitReflectionButton from '../ReflectionCard/SubmitReflectionButton'
+import {TipTapEditor} from '../TipTapEditor/TipTapEditor'
 import getBBox from './getBBox'
 import HTMLReflection from './HTMLReflection'
 import type {ReflectColumnCardInFlight} from './PhaseItemColumn'
