+++ packages/client/components/TeamPrompt/TeamPromptResponseCard.tsx
@@ -24,7 +24,7 @@ import {mergeRefs} from '../../utils/react/mergeRefs'
 import SendClientSideEvent from '../../utils/SendClientSideEvent'
 import Avatar from '../Avatar/Avatar'
 import PlainButton from '../PlainButton/PlainButton'
-import PromptResponseEditor from '../promptResponse/PromptResponseEditor'
+import PromptResponseEditor from './PromptResponseEditor'
 import {ResponseCardDimensions} from './TeamPromptGridDimensions'
 import TeamPromptLastUpdatedTime from './TeamPromptLastUpdatedTime'
 import TeamPromptRepliesAvatarList from './TeamPromptRepliesAvatarList'
