+++ packages/client/hooks/useTipTapCommentEditor.ts
@@ -4,8 +4,8 @@ import {Extension, useEditor} from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
 import {useRef} from 'react'
 import type Atmosphere from '../Atmosphere'
-import {LoomExtension} from '../components/promptResponse/loomExtension'
-import {TiptapLinkExtension} from '../components/promptResponse/TiptapLinkExtension'
+import {LoomExtension} from '../components/TipTapEditor/LoomExtension'
+import {TiptapLinkExtension} from '../components/TipTapEditor/TiptapLinkExtension'
 import {mentionConfig} from '../shared/tiptap/serverTipTapExtensions'
 import {ClearOnSubmit} from '../tiptap/extensions/ClearOnSubmit'
 import {MentionTaskTag} from '../utils/MentionTaskTag'
