+++ packages/client/hooks/useTipTapPageEditor.ts
@@ -17,8 +17,8 @@ import {useRef} from 'react'
 import {readInlineData} from 'relay-runtime'
 import AutoJoiner from 'tiptap-extension-auto-joiner'
 import type {useTipTapPageEditor_viewer$key} from '../__generated__/useTipTapPageEditor_viewer.graphql'
-import {LoomExtension} from '../components/promptResponse/loomExtension'
-import {TiptapLinkExtension} from '../components/promptResponse/TiptapLinkExtension'
+import {LoomExtension} from '../components/TipTapEditor/LoomExtension'
+import {TiptapLinkExtension} from '../components/TipTapEditor/TiptapLinkExtension'
 import {useUploadUserAsset} from '../mutations/useUploadUserAsset'
 import {themeBackgroundColors} from '../shared/themeBackgroundColors'
 import FileBlock from '../tiptap/extensions/fileBlock/FileBlock'
