+++ packages/client/hooks/useTipTapReflectionEditor.ts
@@ -6,8 +6,8 @@ import {Extension, generateText, useEditor} from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
 import {useEffect, useRef, useState} from 'react'
 import type Atmosphere from '../Atmosphere'
-import {LoomExtension} from '../components/promptResponse/loomExtension'
-import {TiptapLinkExtension} from '../components/promptResponse/TiptapLinkExtension'
+import {LoomExtension} from '../components/TipTapEditor/LoomExtension'
+import {TiptapLinkExtension} from '../components/TipTapEditor/TiptapLinkExtension'
 import {isEqualWhenSerialized} from '../shared/isEqualWhenSerialized'
 import {mentionConfig, serverTipTapExtensions} from '../shared/tiptap/serverTipTapExtensions'
 import {ClearOnSubmit} from '../tiptap/extensions/ClearOnSubmit'
