+++ packages/client/modules/outcomeCard/components/OutcomeCard/OutcomeCard.tsx
@@ -8,9 +8,9 @@ import type {AreaEnum, TaskStatusEnum} from '~/__generated__/UpdateTaskMutation.
 import EditingStatus from '~/components/EditingStatus/EditingStatus'
 import {PALETTE} from '~/styles/paletteV3'
 import IntegratedTaskContent from '../../../../components/IntegratedTaskContent'
-import {TipTapEditor} from '../../../../components/promptResponse/TipTapEditor'
 import TaskIntegrationLink from '../../../../components/TaskIntegrationLink'
 import TaskWatermark from '../../../../components/TaskWatermark'
+import {TipTapEditor} from '../../../../components/TipTapEditor/TipTapEditor'
 import useAtmosphere from '../../../../hooks/useAtmosphere'
 import type {UseTaskChild} from '../../../../hooks/useTaskChildFocus'
 import UpdateTaskMutation from '../../../../mutations/UpdateTaskMutation'
