+++ packages/client/shared/tiptap/serverTipTapExtensions.ts
@@ -6,7 +6,7 @@ import {Table, TableCell, TableHeader, TableRow} from '@tiptap/extension-table'
 import {Markdown} from '@tiptap/markdown'
 import StarterKit from '@tiptap/starter-kit'
 import {PageUserMention} from '~/tiptap/extensions/pageUserMention/PageUserMention'
-import {LoomExtension} from '../../components/promptResponse/loomExtension'
+import {LoomExtension} from '../../components/TipTapEditor/LoomExtension'
 import {Database} from '../../tiptap/extensions/database/Database'
 import {UniqueID} from '../../tiptap/extensions/docWithID/UniqueID'
 import {ImageBlockBase} from '../../tiptap/extensions/imageBlock/ImageBlockBase'
