+++ packages/mattermost-plugin/hooks/useTipTapTaskEditor.ts
@@ -2,7 +2,7 @@ import Mention from '@tiptap/extension-mention'
 import {Placeholder} from '@tiptap/extensions'
 import {useEditor} from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
-import {TiptapLinkExtension} from 'parabol-client/components/promptResponse/TiptapLinkExtension'
+import {TiptapLinkExtension} from 'parabol-client/components/TipTapEditor/TiptapLinkExtension'
 import {useTipTapEditorContent} from 'parabol-client/hooks/useTipTapEditorContent'
 import {MentionTaskTag} from 'parabol-client/utils/MentionTaskTag'
 import {tiptapEmojiConfig} from 'parabol-client/utils/tiptapEmojiConfig'
